+++ src/lib/mutations.js
@@ -1,0 +1,3 @@
+export default {
+  createGuest: (root, args) => {}
+}
